refactor logger: move subscription initialization into separate class
