+++ src/modules/logger/CMakeLists.txt
@@ -38,6 +38,7 @@ px4_add_module(
 	COMPILE_FLAGS
 		-Wno-cast-align # TODO: fix and enable
 	SRCS
+		logged_topics.cpp
 		logger.cpp
 		log_writer.cpp
 		log_writer_file.cpp
